+++ CoreInteractiveGraph/src/au/gov/asd/tac/constellation/graph/interaction/InteractiveGraphPluginRegistry.java
@@ -20,7 +20,6 @@ import au.gov.asd.tac.constellation.graph.interaction.plugins.clipboard.CopyToNe
 import au.gov.asd.tac.constellation.graph.interaction.plugins.clipboard.CutToClipboardPlugin;
 import au.gov.asd.tac.constellation.graph.interaction.plugins.clipboard.PasteFromClipboardPlugin;
 import au.gov.asd.tac.constellation.graph.interaction.plugins.clipboard.PasteGraphPlugin;
-import au.gov.asd.tac.constellation.graph.interaction.plugins.clipboard.PasteTextPlugin;
 import au.gov.asd.tac.constellation.graph.interaction.plugins.composite.ContractAllCompositesPlugin;
 import au.gov.asd.tac.constellation.graph.interaction.plugins.composite.CreateCompositeFromSelectionPlugin;
 import au.gov.asd.tac.constellation.graph.interaction.plugins.composite.CreateCompositesFromDominantNodesPlugin;
@@ -76,7 +75,6 @@ public class InteractiveGraphPluginRegistry {
     public static final String FREEFORM_SELECTION = FreeformSelectionPlugin.class.getName();
     public static final String PASTE = PasteFromClipboardPlugin.class.getName();
     public static final String PASTE_GRAPH = PasteGraphPlugin.class.getName();
-    public static final String PASTE_TEXT = PasteTextPlugin.class.getName();
     public static final String POINT_SELECTION = PointSelectionPlugin.class.getName();
     public static final String PREVIOUS_VIEW = PreviousViewPlugin.class.getName();
     public static final String RESET_VIEW = ResetViewPlugin.class.getName();