+++ CoreInteractiveGraph/src/au/gov/asd/tac/constellation/graph/interaction/plugins/clipboard/CutToClipboardPlugin.java
@@ -62,10 +62,10 @@ public final class CutToClipboardPlugin extends SimpleEditPlugin {
                     wg.removeTransaction(id);
                 }
             }
-
-            final String msg = Bundle.MSG_Cut(vxCopied.cardinality(), txCopied.cardinality());
+            
             final StatusDisplayer statusDisplayer = StatusDisplayer.getDefault();
             if (statusDisplayer != null) {
+                final String msg = String.format("Nodes cut: %d; Transactions cut: %d", vxCopied.cardinality(), txCopied.cardinality());
                 statusDisplayer.setStatusText(msg);
             }
         } else {