+++ CoreInteractiveGraph/src/au/gov/asd/tac/constellation/graph/interaction/plugins/composite/DestroyAllCompositesAction.java
@@ -20,7 +20,6 @@ import au.gov.asd.tac.constellation.graph.node.GraphNode;
 import au.gov.asd.tac.constellation.graph.node.plugins.SimplePluginAction;
 import org.openide.awt.ActionID;
 import org.openide.awt.ActionReference;
-import org.openide.awt.ActionReferences;
 import org.openide.awt.ActionRegistration;
 import org.openide.util.NbBundle.Messages;
 
@@ -33,9 +32,7 @@ import org.openide.util.NbBundle.Messages;
 @ActionID(category = "Tools", id = "au.gov.asd.tac.constellation.functionality.composite.DestroyAllCompositesAction")
 @ActionRegistration(displayName = "#CTL_DestroyAllCompositesAction",
         iconBase = "au/gov/asd/tac/constellation/graph/interaction/plugins/composite/resources/destroyCompositeNodes.png")
-@ActionReferences({
-    @ActionReference(path = "Menu/Tools", position = 1000)
-})
+@ActionReference(path = "Menu/Tools", position = 1000)
 @Messages("CTL_DestroyAllCompositesAction=Destroy All Composites")
 public class DestroyAllCompositesAction extends SimplePluginAction {