+++ CoreInteractiveGraph/src/au/gov/asd/tac/constellation/graph/interaction/plugins/zoom/ResetToNZAxisAction.java
@@ -15,7 +15,6 @@
  */
 package au.gov.asd.tac.constellation.graph.interaction.plugins.zoom;
 
-import au.gov.asd.tac.constellation.graph.Graph;
 import au.gov.asd.tac.constellation.graph.interaction.InteractiveGraphPluginRegistry;
 import au.gov.asd.tac.constellation.graph.node.GraphNode;
 import au.gov.asd.tac.constellation.plugins.PluginExecution;
@@ -26,7 +25,6 @@ import javax.swing.Action;
 import javax.swing.Icon;
 import org.openide.awt.ActionID;
 import org.openide.awt.ActionReference;
-import org.openide.awt.ActionReferences;
 import org.openide.awt.ActionRegistration;
 import org.openide.util.NbBundle;
 
@@ -36,9 +34,7 @@ import org.openide.util.NbBundle;
  */
 @ActionID(category = "Display", id = "au.gov.asd.tac.constellation.functionality.zoom.ResetToNZAxisAction")
 @ActionRegistration(displayName = "#CTL_ResetToNZAxisAction", iconBase = "au/gov/asd/tac/constellation/graph/interaction/plugins/zoom/resources/axis_z_negative.png", surviveFocusChange = true)
-@ActionReferences({
-    @ActionReference(path = "Menu/Display/Reset View by Axis", position = 600)
-})
+@ActionReference(path = "Menu/Display/Reset View by Axis", position = 600)
 @NbBundle.Messages("CTL_ResetToNZAxisAction=-Z Axis")
 public class ResetToNZAxisAction extends AbstractAction {
 
@@ -46,17 +42,16 @@ public class ResetToNZAxisAction extends AbstractAction {
     private static final Icon AXIS_NZ_ICON = UserInterfaceIconProvider.AXIS_Z_NEGATIVE.buildIcon(16);
 
     public ResetToNZAxisAction(final GraphNode context) {
-        putValue(Action.SMALL_ICON, AXIS_NZ_ICON);
         this.context = context;
+        putValue(Action.SMALL_ICON, AXIS_NZ_ICON);
     }
 
     @Override
-    public void actionPerformed(ActionEvent e) {
-        final Graph graph = context.getGraph();
+    public void actionPerformed(final ActionEvent e) {
         PluginExecution.withPlugin(InteractiveGraphPluginRegistry.RESET_VIEW)
                 .withParameter(ResetViewPlugin.AXIS_PARAMETER_ID, "z")
                 .withParameter(ResetViewPlugin.NEGATIVE_PARAMETER_ID, true)
                 .withParameter(ResetViewPlugin.SIGNIFICANT_PARAMETER_ID, true)
-                .executeLater(graph);
+                .executeLater(context.getGraph());
     }
 }