+++ CoreInteractiveGraph/src/au/gov/asd/tac/constellation/graph/interaction/plugins/zoom/ZoomOutAction.java
@@ -17,9 +17,7 @@ package au.gov.asd.tac.constellation.graph.interaction.plugins.zoom;
 
 import au.gov.asd.tac.constellation.graph.interaction.InteractiveGraphPluginRegistry;
 import au.gov.asd.tac.constellation.graph.node.GraphNode;
-import au.gov.asd.tac.constellation.plugins.PluginExecution;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
+import au.gov.asd.tac.constellation.graph.node.plugins.SimplePluginAction;
 import org.openide.awt.ActionID;
 import org.openide.awt.ActionReference;
 import org.openide.awt.ActionReferences;
@@ -40,21 +38,9 @@ import org.openide.util.NbBundle;
     @ActionReference(path = "Shortcuts", name = "D-SUBTRACT")
 })
 @NbBundle.Messages("CTL_ZoomOut=Zoom Out")
-public final class ZoomOutAction implements ActionListener {
-
-    private final GraphNode context;
-
-    /**
-     * Construct a new ZoomOutAction.
-     *
-     * @param context GraphNode.
-     */
+public final class ZoomOutAction extends SimplePluginAction {
+    
     public ZoomOutAction(final GraphNode context) {
-        this.context = context;
-    }
-
-    @Override
-    public void actionPerformed(final ActionEvent e) {
-        PluginExecution.withPlugin(InteractiveGraphPluginRegistry.ZOOM_OUT).executeLater(context.getGraph());
+        super(context, InteractiveGraphPluginRegistry.ZOOM_OUT);
     }
 }